refactor: extract constants and utilities
