+++ src/modules/Dashboard/features/AccountSettings/MFARecoveryCodes/index.ts
@@ -2,3 +2,4 @@ export { MFARecoveryCodes } from './MFARecoveryCodes';
 export type { MFARecoveryCodesProps } from './MFARecoveryCodes.types';
 export { useRecoveryCodesDownload } from './useRecoveryCodesDownload';
 export * from './MFARecoveryCodes.utils';
+export * from './fileDownload.utils';
