+++ src/modules/Dashboard/features/AccountSettings/RemoveMFA/index.ts
@@ -3,3 +3,6 @@ export { RemoveMFAConfirmation } from './RemoveMFAConfirmation';
 export { useRemoveMFA } from './useRemoveMFA';
 export type { RemoveMFAConfirmationProps } from './RemoveMFAConfirmation.types';
 export type { RemoveMFAProps } from './RemoveMFA.types';
+export { MFA_DISABLE_ERROR_MESSAGES } from './RemoveMFA.constants';
+export { getErrorMessage } from './RemoveMFA.utils';
+export type { ErrorResponse } from './RemoveMFA.utils';
