+++ src/modules/Dashboard/features/AccountSettings/shared/index.ts
@@ -7,3 +7,6 @@ export { useMFAInputHandler } from './useMFAInputHandler';
 export type { ConfirmIdentityVerificationCodeProps } from './ConfirmIdentityVerificationCode.types';
 export type { ConfirmIdentityRecoveryCodeProps } from './ConfirmIdentityRecoveryCode.types';
 export type { MFAVerificationFormProps } from './MFAVerificationForm.types';
+
+// Shared styled components for MFA dialogs
+export * from './MFADialog.styles';
